Also build a 64-bit exe version of the 32-bit courgette utility + library to support larger input sizes.
BUG=63793
TEST=None


Review URL: http://codereview.chromium.org/5096007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@69779 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 104a608acbf0fe57728f95d21c59ed25f61b0482

+++ courgette/adjustment_method.cc
@@ -524,7 +524,7 @@ class AssignmentProblem {
   Node* MakeRootNode(const Trace& trace) {
     Node* node = new Node(NULL, NULL);
     all_nodes_.push_back(node);
-    for (size_t i = 0;  i < trace.size();  ++i) {
+    for (uint32 i = 0;  i < trace.size();  ++i) {
       ++node->count_;
       node->places_.push_back(i);
     }
@@ -635,7 +635,8 @@ class GraphAdjuster : public AdjustmentMethod {
   }
 
   void ReferenceLabel(Trace* trace, Label* label, bool is_model) {
-    trace->push_back(MakeLabelInfo(label, is_model, trace->size()));
+    trace->push_back(MakeLabelInfo(label, is_model,
+                                   static_cast<uint32>(trace->size())));
   }
 
   LabelInfo* MakeLabelInfo(Label* label, bool is_model, uint32 position) {
